Profile readme update
If you are a professional company, my current or future employer. This is a statement that in no way does this relate to my professional career or views, what is expressed here is purely my personal, moral, and ethical opinion about modern digital ethics and how corrupted the human species is becoming. pls no haterino